Fixed svfft
