+++ src/LinearAlgebra/helper.jl
@@ -176,3 +176,47 @@ end
 
 interlace{T}(a::Vector{T},b::Vector{T})=interlace(Vector{T}[a,b])
 
+
+
+
+## svfft
+
+##FFT That interlaces coefficients
+function svfft(v::Vector)
+        n=length(v)
+        v=FFTW.fft(v)./n
+        if mod(n,2) == 0
+            ind=div(n,2)
+            v=alternatesign!(v)
+            interlace(v[1:ind],
+                      flipdim(v[ind+1:end],1))
+        elseif mod(n,4)==3
+            ind=div(n+1,2)
+            interlace(alternatesign!(v[1:ind]),
+                      -flipdim(alternatesign!(v[ind+1:end]),1))
+        else #mod(length(v),4)==1
+            ind=div(n+1,2)
+            interlace(alternatesign!(v[1:ind]),
+                      flipdim(alternatesign!(v[ind+1:end]),1))
+        end
+end
+
+
+
+
+
+function isvfft(sv::Vector)
+        n=length(sv)
+
+        if mod(n,2) == 0
+            v=alternatesign!([sv[1:2:end];flipdim(sv[2:2:end],1)])
+        elseif mod(n,4)==3
+            v=[alternatesign!(sv[1:2:end]);
+                -alternatesign!(flipdim(sv[2:2:end],1))]
+        else #mod(length(v),4)==1
+            v=[alternatesign!(sv[1:2:end]);
+                    alternatesign!(flipdim(sv[2:2:end],1))]
+        end
+
+        FFTW.ifft(n*v)
+end
